Move scripts into subcategory folders
